inference_operator superseeded by inference_manager; doc; version bump
