+++ docs/api/inference.md
@@ -1,5 +1,18 @@
 # inference modules
 
+This section contains documentation for the core inference functionality in InfOCF.
+
+## Main Interface
+
+::: inference.inference_manager
+    options:
+      show_source: false
+      filters:
+        - "!^_"
+      members_order: source
+
+## Core Components
+
 ::: inference.inference
     options:
       show_source: false
@@ -13,6 +26,13 @@
         - "!^_"
 
 ::: inference.belief_base
+    options:
+      show_source: false
+      filters:
+        - "!^_"
+
+
+::: inference.queries
     options:
       show_source: false
       filters:
